Split gitoid crate into modules.
This commit breaks up the single-file `gitoid` crate into multiple
modules, organized basically per-type. This helps to clarify
imports and interfaces, and should (hopefully) make the code more
navigable, particularly as it continues to grow.
